refactoring test classes
